Change back to nalgebra. Not sure why this fixes the test for integration. Added ramp.

+++ src/constant.rs
@@ -1,28 +1,28 @@
 //! Mathematical and physical constants.
 
 /// Reduced plank constant in SI units
-pub const HBAR: f32 = 1.0545718e-34;
+pub const HBAR: f64 = 1.0545718e-34;
 
 /// Gravitational acceleration in SI units
-pub const GC: f32 = 9.80665;
+pub const GC: f64 = 9.80665;
 
 /// Mathematical constant exp(1)
-pub const EXP: f32 = std::f32::consts::E;
+pub const EXP: f64 = std::f64::consts::E;
 
 /// Mathematica constant pi
-pub const PI: f32 = std::f32::consts::PI;
+pub const PI: f64 = std::f64::consts::PI;
 
 // The Bohr magneton, defined in SI units of Joules/Tesla.
-pub const BOHRMAG: f32 = 9.274e-24;
+pub const BOHRMAG: f64 = 9.274e-24;
 
 /// Boltzmann constant in SI units
-pub const BOLTZCONST: f32 = 1.38e-23;
+pub const BOLTZCONST: f64 = 1.38e-23;
 
 /// The value of 1 Atomic Mass Unit (amu) in SI units of kg.
-pub const AMU: f32 = 1.6605e-27;
+pub const AMU: f64 = 1.6605e-27;
 
 /// Speed of light in SI units of m/s
-pub const C: f32 = 299297458.0;
+pub const C: f64 = 299297458.0;
 
 /// Sqrt of 2
 pub const SQRT2: f64 = std::f64::consts::SQRT_2;
